% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{figheight}
\alias{figheight}
\title{Calculate Figure Height}
\usage{
figheight(
  obj,
  trim5 = 0,
  trim3 = 0,
  width = 100,
  pixelsperrow = 200,
  showtrim = FALSE
)
}
\description{
Calculate Figure Height
}
