% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pickchromatogram}
\alias{pickchromatogram}
\title{Pick Chromatogram}
\usage{
pickchromatogram(
  obj,
  trim5 = 0,
  trim3 = 0,
  width = 100,
  pixelsperrow = 200,
  showtrim = FALSE,
  inputloc,
  selectedGenomeVersion,
  custom_seq = NULL
)
}
\description{
Pick Chromatogram
}
