% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_shelper}
\alias{run_shelper}
\title{Run the Shelper Shiny App}
\usage{
run_shelper(port = 3838, launch.browser = TRUE)
}
\arguments{
\item{port}{The port to run the application on. Default is 3838.}

\item{launch.browser}{Whether to launch the browser. Default is TRUE.}
}
\description{
Run the Shelper Shiny App
}
