% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_shelper_config}
\alias{set_shelper_config}
\title{Set Shelper Configuration}
\usage{
set_shelper_config(
  hg19 = NULL,
  hg38 = NULL,
  blastn = NULL,
  blast_db = NULL,
  temp_path = NULL
)
}
\arguments{
\item{hg19}{Path to hg19 fasta}

\item{hg38}{Path to hg38 fasta}

\item{blastn}{Path to blastn executable}

\item{blast_db}{Path to blast database}

\item{temp_path}{Path for temporary files}
}
\description{
Set Shelper Configuration
}
